refactored log event code from UacProcessor to new class
