+++ src/test/java/uk/gov/ons/census/casesvc/service/EventProcessorTest.java
@@ -18,6 +18,7 @@ import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.ons.census.casesvc.logging.EventLogger;
 import uk.gov.ons.census.casesvc.model.dto.CreateCaseSample;
 import uk.gov.ons.census.casesvc.model.dto.EventDTO;
 import uk.gov.ons.census.casesvc.model.dto.EventTypeDTO;
@@ -38,6 +39,8 @@ public class EventProcessorTest {
   @Mock UacProcessor uacProcessor;
 
   @Mock EventRepository eventRepository;
+
+  @Mock EventLogger eventLogger;
 
   @InjectMocks EventProcessor underTest;
 
@@ -62,7 +65,7 @@ public class EventProcessorTest {
     verify(uacProcessor).saveUacQidLink(eq(caze), eq(1));
     verify(uacProcessor).emitUacUpdatedEvent(uacQidLink, caze);
     verify(caseProcessor).emitCaseCreatedEvent(caze);
-    verify(uacProcessor, times(2))
+    verify(eventLogger, times(2))
         .logEvent(eq(uacQidLink), any(String.class), any(EventType.class), any(PayloadDTO.class));
   }
 
@@ -89,7 +92,7 @@ public class EventProcessorTest {
     verify(uacProcessor, times(1)).saveUacQidLink(eq(caze), eq(2));
     verify(uacProcessor, times(2)).emitUacUpdatedEvent(uacQidLink, caze);
     verify(caseProcessor).emitCaseCreatedEvent(caze);
-    verify(uacProcessor, times(3))
+    verify(eventLogger, times(3))
         .logEvent(eq(uacQidLink), any(String.class), any(EventType.class), any(PayloadDTO.class));
   }
