+++ src/test/java/uk/gov/ons/census/casesvc/service/ReceiptProcessorTest.java
@@ -1,6 +1,11 @@
 package uk.gov.ons.census.casesvc.service;
 
-import static org.mockito.Mockito.*;
+import static org.mockito.Mockito.any;
+import static org.mockito.Mockito.anyBoolean;
+import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.times;
+import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.when;
 import static uk.gov.ons.census.casesvc.service.ReceiptProcessor.QID_RECEIPTED;
 
 import java.time.OffsetDateTime;
@@ -10,6 +15,11 @@ import java.util.Optional;
 import java.util.UUID;
 import org.jeasy.random.EasyRandom;
 import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.MockitoJUnitRunner;
+import uk.gov.ons.census.casesvc.logging.EventLogger;
 import uk.gov.ons.census.casesvc.model.dto.PayloadDTO;
 import uk.gov.ons.census.casesvc.model.dto.Receipt;
 import uk.gov.ons.census.casesvc.model.entity.Case;
@@ -18,11 +28,24 @@ import uk.gov.ons.census.casesvc.model.entity.UacQidLink;
 import uk.gov.ons.census.casesvc.model.repository.CaseRepository;
 import uk.gov.ons.census.casesvc.model.repository.UacQidLinkRepository;
 
+@RunWith(MockitoJUnitRunner.class)
 public class ReceiptProcessorTest {
   private static final UUID TEST_CASE_ID = UUID.randomUUID();
   private static final String TEST_QUESTIONNAIRE_ID = "123";
   private static final String TEST_QID = "test_qid";
   private static final String TEST_UAC = "test_uac";
+
+  @Mock private CaseProcessor caseProcessor;
+
+  @Mock private UacQidLinkRepository uacQidLinkRepository;
+
+  @Mock private CaseRepository caseRepository;
+
+  @Mock private UacProcessor uacProcessor;
+
+  @Mock private EventLogger eventLogger;
+
+  @InjectMocks UacProcessor underTest;
 
   @Test
   public void testGoodReceipt() {
@@ -55,12 +78,13 @@ public class ReceiptProcessorTest {
     receipt.setResponseDateTime(expectedReceiptDateTime);
 
     ReceiptProcessor receiptProcessor =
-        new ReceiptProcessor(caseProcessor, uacQidLinkRepository, caseRepository, uacProcessor);
+        new ReceiptProcessor(
+            caseProcessor, uacQidLinkRepository, caseRepository, uacProcessor, eventLogger);
     receiptProcessor.processReceipt(receipt, headers);
 
     // then
     verify(uacProcessor, times(1)).emitUacUpdatedEvent(expectedUacQidLink, expectedCase, false);
-    verify(uacProcessor, times(1))
+    verify(eventLogger, times(1))
         .logReceiptEvent(
             expectedUacQidLink,
             QID_RECEIPTED,
@@ -83,7 +107,8 @@ public class ReceiptProcessorTest {
     receipt.setQuestionnaire_Id(TEST_QUESTIONNAIRE_ID);
 
     ReceiptProcessor receiptProcessor =
-        new ReceiptProcessor(caseProcessor, uacQidLinkRepository, caseRepository, uacProcessor);
+        new ReceiptProcessor(
+            caseProcessor, uacQidLinkRepository, caseRepository, uacProcessor, eventLogger);
     receiptProcessor.processReceipt(receipt, createTestHeaders());
 
     // Then
